+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 ## Alpha
+### v0.2
+- Change status placeholders
+- Fix enemy shield bug
 ### v0.1
 - Game is playable and can be finished
 - It's first version that was distributed